web/requirements.txt
Flask
Flask-RESTful
pymongo
bcrypt
numpy
tensorflow
requests
keras
pillow
